requirements.txt
numpy
pillow
opencv-python
pyyaml
scikit-learn
timm
torch
torchvision
tqdm
